QuickDropApp/requirements.txt
streamlit>=1.36
sqlalchemy>=2.0
psycopg[binary]>=3.1
pandas>=2.2
python-dotenv>=1.0
plotly>=5.22
python-dateutil>=2.9
